Api to track guarantee details in sold orders

+++ src/main/java/com/project/JewelryMS/entity/ProductSell.java
@@ -39,7 +39,7 @@ public class ProductSell {
     private List<ProductSell_Promotion> productSellPromotions;
 
     @Column(name = "chi")
-    private Integer chi;
+    private Float chi;
 
     @Column(name = "cost")
     private float cost;
