+++ src/main/java/com/project/JewelryMS/model/ProductBuy/CalculatePBRequest.java
@@ -11,6 +11,6 @@ import org.springframework.web.multipart.MultipartFile;
 public class CalculatePBRequest {
     private String metalType;
     private String gemstoneType;
-    private Integer metalWeight;
+    private Float metalWeight;
     private Float gemstoneWeight;
 }
